Xpetra MatrixMatrix: ETI
