+++ packages/xpetra/sup/Utils/ExplicitInstantiation/ETI_SC_LO_GO_NO_classes.cmake
@@ -1,3 +1,5 @@
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::CrsMatrixWrap)
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::IO)
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::MatrixUtils)
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::MatrixMatrix)
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::Helpers)
