+++ packages/xpetra/sup/Utils/Xpetra_MatrixUtils_decl.hpp
@@ -59,8 +59,7 @@
 #include "Xpetra_MatrixFactory.hpp"
 #include "Xpetra_BlockedCrsMatrix.hpp"
 #include "Xpetra_MatrixMatrix.hpp"
-
-#include "Xpetra_IO.hpp"
+#include "Xpetra_Helpers.hpp"
 
 #ifdef HAVE_XPETRA_TPETRA
 #include "Xpetra_TpetraMultiVector.hpp"
